backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
pydantic-settings
python-multipart
httpx
jinja2
pytest
pytest-asyncio
python-jose[cryptography]
passlib[bcrypt]
google-auth
google-auth-oauthlib
google-auth-httplib2
tenacity
google-api-core
deep-translator
gTTS
